backend/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-socketio==5.11.0
aiohttp==3.9.1
aiohttp-cors==0.7.0